requirements.txt
Django~=3.2.12
djangorestframework~=3.12.4
psycopg2~=2.9.3
django-storages~=1.9.1
azure-common~=1.1.28
azure-storage-blob~=2.1.0
azure-storage-common~=2.1.0